TST: Add tests for pivot_table
