+++ tests/test_pandas.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
 
+import datetime as dt
 import random
 from typing import (
     TYPE_CHECKING,
@@ -1102,4 +1103,196 @@ def test_merge_asof() -> None:
             pd.DataFrame,
         ),
         pd.DataFrame,
+    )
+
+
+def test_pivot_extended() -> None:
+    df = pd.DataFrame(
+        data={
+            "col1": ["first", "second", "third", "fourth"],
+            "col2": [50, 70, 56, 111],
+            "col3": ["A", "B", "B", "A"],
+            "col4": [100, 102, 500, 600],
+            ("col5",): ["E", "F", "G", "H"],
+            ("col6", 6): ["apple", "banana", "cherry", "date"],
+            ("col7", "other"): ["apple", "banana", "cherry", "date"],
+            dt.date(2000, 1, 1): ["E", "F", "G", "H"],
+            dt.datetime(2001, 1, 1, 12): ["E", "F", "G", "H"],
+            dt.timedelta(7): ["E", "F", "G", "H"],
+            True: ["E", "F", "G", "H"],
+            9: ["E", "F", "G", "H"],
+            10.0: ["E", "F", "G", "H"],
+            (11.0 + 1j): ["E", "F", "G", "H"],
+            pd.Timestamp(2002, 1, 1): ["E", "F", "G", "H"],
+            pd.Timedelta(1, "D"): ["E", "F", "G", "H"],
+        }
+    )
+    check(
+        assert_type(
+            pd.pivot(df, index="col1", columns="col3", values="col2"), pd.DataFrame
+        ),
+        pd.DataFrame,
+    )
+    check(
+        assert_type(
+            pd.pivot(df, index=[("col5",)], columns=[("col6", 6)], values="col2"),
+            pd.DataFrame,
+        ),
+        pd.DataFrame,
+    )
+    check(
+        assert_type(
+            pd.pivot(
+                df, index=[("col5",)], columns=[("col6", 6)], values=[("col7", "other")]
+            ),
+            pd.DataFrame,
+        ),
+        pd.DataFrame,
+    )
+    check(
+        assert_type(
+            pd.pivot(df, index=dt.date(2000, 1, 1), columns="col3", values="col2"),
+            pd.DataFrame,
+        ),
+        pd.DataFrame,
+    )
+    check(
+        assert_type(
+            pd.pivot(
+                df, index=dt.datetime(2001, 1, 1, 12), columns="col3", values="col2"
+            ),
+            pd.DataFrame,
+        ),
+        pd.DataFrame,
+    )
+    check(
+        assert_type(
+            pd.pivot(df, index=dt.timedelta(7), columns="col3", values="col2"),
+            pd.DataFrame,
+        ),
+        pd.DataFrame,
+    )
+    check(
+        assert_type(
+            pd.pivot(df, index=True, columns="col3", values="col2"), pd.DataFrame
+        ),
+        pd.DataFrame,
+    )
+    check(
+        assert_type(pd.pivot(df, index=9, columns="col3", values="col2"), pd.DataFrame),
+        pd.DataFrame,
+    )
+    check(
+        assert_type(
+            pd.pivot(df, index=10.0, columns="col3", values="col2"), pd.DataFrame
+        ),
+        pd.DataFrame,
+    )
+    check(
+        assert_type(
+            pd.pivot(df, index=(11.0 + 1j), columns="col3", values="col2"), pd.DataFrame
+        ),
+        pd.DataFrame,
+    )
+    check(
+        assert_type(
+            pd.pivot(df, index=pd.Timestamp(2002, 1, 1), columns="col3", values="col2"),
+            pd.DataFrame,
+        ),
+        pd.DataFrame,
+    )
+    check(
+        assert_type(
+            pd.pivot(df, index=pd.Timedelta(1, "D"), columns="col3", values="col2"),
+            pd.DataFrame,
+        ),
+        pd.DataFrame,
+    )
+
+
+def test_pivot_table() -> None:
+    df = pd.DataFrame(
+        {
+            "A": ["foo", "foo", "foo", "foo", "foo", "bar", "bar", "bar", "bar"],
+            "B": ["one", "one", "one", "two", "two", "one", "one", "two", "two"],
+            "C": [
+                "small",
+                "large",
+                "large",
+                "small",
+                "small",
+                "large",
+                "small",
+                "small",
+                "large",
+            ],
+            "D": [1, 2, 2, 3, 3, 4, 5, 6, 7],
+            "E": [2, 4, 5, 5, 6, 6, 8, 9, 9],
+            ("col5",): ["foo", "foo", "foo", "foo", "foo", "bar", "bar", "bar", "bar"],
+            ("col6", 6): [
+                "one",
+                "one",
+                "one",
+                "two",
+                "two",
+                "one",
+                "one",
+                "two",
+                "two",
+            ],
+            (7, "seven"): [
+                "small",
+                "large",
+                "large",
+                "small",
+                "small",
+                "large",
+                "small",
+                "small",
+                "large",
+            ],
+        }
+    )
+    pd.pivot_table(df, values="D", index=["A", "B"], columns=["C"], aggfunc=np.sum)
+    pd.pivot_table(df, values="D", index=["A", "B"], columns="C", aggfunc=np.sum)
+    pd.pivot_table(
+        df, values="D", index=["A", "B"], columns=[(7, "seven")], aggfunc=np.sum
+    )
+    pd.pivot_table(
+        df,
+        values="D",
+        index=[("col5",), ("col6", 6)],
+        columns=[(7, "seven")],
+        aggfunc=np.sum,
+    )
+    pd.pivot_table(df, values="D", index=["A", "B"], columns=["C"], aggfunc="sum")
+
+    def f(x: pd.Series) -> float:
+        return x.sum()
+
+    pd.pivot_table(df, values="D", index=["A", "B"], columns=["C"], aggfunc=f)
+    pd.pivot_table(df, values="D", index=["A", "B"], columns=["C"], aggfunc={"D": f})
+    pd.pivot_table(
+        df, values="D", index=["A", "B"], columns=["C"], aggfunc=[f, np.sum, "sum"]
+    )
+    pd.pivot_table(
+        df,
+        values="D",
+        index=["A", "B"],
+        columns=["C"],
+        aggfunc=np.sum,
+        margins=True,
+        margins_name="Total",
+    )
+    pd.pivot_table(
+        df, values="D", index=["A", "B"], columns=["C"], aggfunc=np.sum, dropna=True
+    )
+    pd.pivot_table(
+        df, values="D", index=["A", "B"], columns=["C"], aggfunc=np.sum, dropna=True
+    )
+    pd.pivot_table(
+        df, values="D", index=["A", "B"], columns=["C"], aggfunc=np.sum, observed=True
+    )
+    pd.pivot_table(
+        df, values="D", index=["A", "B"], columns=["C"], aggfunc=np.sum, sort=False
     )
